Refactor cog blocks into a separate module.
